feat: add multi-tenancy docs for self-managed (#2727)
* Document multi-tenancy for Zeebe gRPC (#2696)

* feat(apis-tools): document tenant-related grpc properties

Document the tenantId and tenantIds properties in all of the
gRPC request and response messages that were made tenant-aware.

In some gRPC requests the tenantId property is not required as
the tenant ownership of the data is already clear within Zeebe.

* feat(apis-tools): document tenant-related grpc errors

Document the multi-tenancy-related errors that may be produced
by tenant-aware gRPC requests.

* doc(api-tools): document multi-tenancy limiations of zeebe grpc spec

Document the limitations of multi-tenancy in the Zeebe gRPC
calls. Signals are not supported for non-<default> tenants,
and resource deletion is not tenant-aware, meaning that it
can delete any resource no matter the tenant ownership.

* doc(api-tools): multi-tenancy for deprecated rpc

Document that for the deprecated DeployProcess RPC, when
multi-tenancy is enabled, any deployments are owned by
the <default> tenant. Deploying to other tenants through
this RPC is not possible since it's deprecated and no
new behaviors are added.

* fix(grpc): correct multi-tenancy support note for deletion

Correct the note on multi-tenancy support for resource deletion in
the gRPC documentation. Resource deletion is only supported for the
<default> tenant.

* docs(identity): introduce Identity specific documentation relating to multi tenancy (#2713)

* docs(identity): introduce Identity specific documentation relating to multi tenancy

* style(formatting): technical review

* docs(identity): PR comments

---------

Co-authored-by: Christina Ausley <christina.ausley@camunda.com>

* Add multi-tenancy config docs for Connectors

* chore(multiTenancy): document C8SM multi tenancy feature for Optimize

related to OPT-7250

* feat: add multi tenancy section to Job Worker


This commit adds a section about multi-tenancy in Job workers and how to configure them. It also includes some small code examples to guide users during implementation.

Co-authored-by: Christina Ausley <christina.ausley@camunda.com>

* chore(version-8.3): backport multi-tenancy changes

* revert: remove mt sidebar changes

The conceptual multi-tenancy docs haven't been added yet,
so we shouldn't link to them from the sidebar.

* docs(MT:tasklist): document MT-related changes for Tasklist (#2730)

* docs(MT:tasklist): document MT-related changes for Tasklist

* docs(MT:tasklist): port documentation for v`8.3` also

* technical review

---------

Co-authored-by: Christina Ausley <christina.ausley@camunda.com>

* Document multi-tenancy support for Zeebe (#2734)

* doc(go-client): document multi-tenancy support for the go client

Document that multi-tenancy is currently not supported for the
go client.

* doc(gateway): document multi-tenancy configuration for zeebe

Document the configuration properties needed to enable
multi-tenancy in Zeebe.

* doc(update-guide): multi-tenancy adjustments on migration

Document the impact of multi-tenancy when migrating from older versions of Zebe
.

* style(formatting): technical review

* refactor: implement review hints

* Add conceptual docs for multi-tenancy (#2735)

* feat(self-managed): add conceptual docs on multi-tenancy

* refactor: implement review hints

---------

Co-authored-by: Christina Ausley <christina.ausley@camunda.com>
Co-authored-by: christinaausley <84338309+christinaausley@users.noreply.github.com>

---------

Co-authored-by: Nikola Koevski <nikola.koevski@camunda.com>
Co-authored-by: Ben Sheppard <Ben-Sheppard@users.noreply.github.com>
Co-authored-by: Christina Ausley <christina.ausley@camunda.com>
Co-authored-by: Dmitriy Melnychuk <dmitriy.melnychuk@capgemini.com>
Co-authored-by: sbuettner <simon.buettner@camunda.com>
Co-authored-by: PHWaechtler <21290204+PHWaechtler@users.noreply.github.com>
Co-authored-by: Remco <remco@westerhoud.nl>
Co-authored-by: Dmitriy Melnychuk <123628255+ce-dmelnych@users.noreply.github.com>
Co-authored-by: christinaausley <84338309+christinaausley@users.noreply.github.com>