Small improvements to configuring multi-tenancy in zeebe gateway (#2761)
* docs(sm): use correct env var

As these are the configurations of the zeebe gateway, the environment
variable should be ZEEBE_GATEWAY_* not ZEEBE_BROKER_GATEWAY_*. Using
ZEEBE_BROKER_GATEWAY_* only makes sense when the gateway is used
embedded inside a broker. See ZEEBE_BROKER_GATEWAY_ENABLE.

I noticed that this is not entirely clear from the specific page, so
I've added a section to clarify this as well and backported it to 8.2
as it holds true there as well. Note that this page did not exist for
the 8.1 docs.

* docs(sm): adjust admonition

The caution admonition was formatted wrongly causing wrong rendering. We
can resolve this with some additional new lines inside the admonition.

* docs(sm): define multi-tenancy config dependencies

In order to configure multi-tenancy in the zeebe gateway, the
authentication mode must be set to identity. This was not clear in the
previous docs.