docs(concepts): describe instance migration basics
Adds some descriptions of basic use cases for process instance migration
that allow users to understand what they can use the feature for and
what not.

I've added several tips to help avoid the quircks of this feature. Note
that this is just the first itaration of this feature and some of these
workarounds will eventually become available as feature options when
migrating.

We still need to add examples using Operate of process models and active
process instances (before and after migration).