Process instance migration concept documentation (#3000)
* docs(concepts): add process-instance-migration page

Introduces a new page for process instance migration, and adds it to the
sidebar.

Process instance migration can be used to change the process definition
of a running process instance. This is useful when the process
definition needs to be changed due to bugs or new requirements.

* docs(concepts): add an initial migration concept

This adds a first text for process instance migration, focussing on the
limitations, and special behavior of expressions, input mappings, and
event subscriptions.

* docs(concepts): describe instance migration basics

Adds some descriptions of basic use cases for process instance migration
that allow users to understand what they can use the feature for and
what not.

I've added several tips to help avoid the quircks of this feature. Note
that this is just the first itaration of this feature and some of these
workarounds will eventually become available as feature options when
migrating.

We still need to add examples using Operate of process models and active
process instances (before and after migration).

* docs(concepts): mark as experimental feature

While we're discussing how to resolve a high severity bug, the feature
is only available by fully pausing all interactions with the process
instance under migration.

We might resolve the bug before the 8.4 release, in which case this
admonition can be removed.

See https://github.com/camunda/zeebe/issues/15129

* fix(concepts): cleanup links

* style(formatting): technical review

* docs(concepts): add basic migration screenshots

* fix: shrink image size

* docs(concepts): add screenshots active migration

* docs(concepts): clarify use case for migration

Provides more insights into when migration is a useful feature.

* fix: remove typo and provide more clarity

* fix: clarify migration of inactive parts

They can't migrate inactive elements, only the flow to inactive parts.

* docs(components): refer to operate user guide

* docs(concepts): add reference to full error codes

* docs: remove comments

* Update docs/components/concepts/process-instance-migration.md

Co-authored-by: berkaycanbc <124141078+berkaycanbc@users.noreply.github.com>

---------

Co-authored-by: Christina Ausley <christina.ausley@camunda.com>
Co-authored-by: christinaausley <84338309+christinaausley@users.noreply.github.com>
Co-authored-by: berkaycanbc <124141078+berkaycanbc@users.noreply.github.com>