docs(concepts): rm reverted event migration section (#3144)
Process instance migration can no longer deal with event subscriptions.
While this is documented in the limitations with some specific
scenarios, it should also be clear in the rest of the concept
documentation. So, this removes the section discussing event
subscriptions because it implies behavior that is not available.

I've also clarified the case where an active service task receives a
boundary event in the target process definition which is also not
supported at this time. It could be implied from the rule above it, but
I'd rather have the limitations be explicit.