feat: Rename Zeebe gPRC page title (Gateway service) (#3561)
* fix: Link to Zeebe gateway protobuf file

The module of the protobuf file was moved. Change the link to the new directory.

* feat: Rename title of Zeebe Gateway Service

The "Gateway service" page contains Zeebe's available RPCs. Technically, the name "Gateway service" is correct because the service is called "Gateway" in the protobuf file. However, users might not expect Zeebe's gRPC API under this name.

Rename the page to "Zeebe API RPCs" and link it from the overview page.