Compensation events (#3494)
* docs(compensation): add compensation events coverage

* docs(compensation): add compesation events

The documentation cover the following points:

How to trigger compensation
Trigger compensation for the scope vs. for a given activity
Trigger compensation from an event subprocess
How to compensate an activity with a compensation handler
Under which conditions is a compensation handler invoked and how often
How compensation works for subprocesses
How compensation works with multi-instance
No compensation for child process instances
When is a compensation handler interrupted
Explain the compensation order (i.e. all compensation handlers at once) and how to enforce a specific order

* docs(compensation): add compesation handler marker

* docs(compensation): apply PR suggestions

_ Moved compensation handlers into the related marker documentation
_ Replace zeebe-play gif with operate gif
_ Rewording of some phrases

* style(formatting): standard review

* docs(compensation): apply PR suggestions

_ Removed gifs in favor of static images
_ Rewording
_ Fix broken links

* docs(compensation): Link to BPMN compensation events page

* docs(compensation): Improve compensation handlers page

* docs(compensation): Improve compensation events page

* style(formatting): technical review

* docs(compensation): Improve call activity handler description

---------

Co-authored-by: Christina Ausley <christina.ausley@camunda.com>
Co-authored-by: Philipp Ossler <philipp.ossler@gmail.com>
Co-authored-by: christinaausley <84338309+christinaausley@users.noreply.github.com>