docs: refer to c8 instance migration concept
This admonition was outdated, as C8 does support instance migration.
This does not further adjust this text to fix the problem fast. We can
adjust the text for C8 usage later.

The feature has been available since 8.4 in very limited form, and has
slowly been growing in capabilities. In 8.6.0-alpha3 there are still
limitations, but I don't we have to mention them here. The limitations
are documented on the concept page and are thus versioned.