feat: document zeebe gateway rest api middleware
Document how custom REST API middleware, i.e. jakarta
servlet filters, can be added to the Zeebe Gateway.

REST API filters can be added to the Zeebe Gateway with
the same mechanism used for gRPC interceptors. As a
result, the documentaion is very similar, but may
diverge as the Camunda 8 REST API develops futher.
This is why there are two separate documentation
sections for filters and interceptors even though
the content is very similar.