Document REST API middleware (#4130)
* feat: remove notes on rest api limitations

With the addition of REST API middleware in the
Zeebe Gateway, some limitations on possible use
cases of the Zeebe REST API are lifted. The docs
should reflect these changes, and as a result
some (warning) notes are removed.

* feat: document rest api middleware support for the java client

Document how REST API middleware can be provided to the
Zeebe Java client by using Apache HttpClient AsyncExecChainHandlers.

The new docs expand on the existing documentation on
gRPC ClientInterceptors for the Zeebe Java client.

Currently, there isn't a separate Camunda docs page
for client middleware, as Zeebe is client agnostic,
i.e. Zeebe provides (gRPC and OpenAPI) specifications
that can be used for implementing Zeebe/Camunda clients.
As such, the Camunda docs can only provide implementation
specifics for middleware in Zeebe clients that are officially
supported.

* feat: document zeebe gateway rest api middleware

Document how custom REST API middleware, i.e. jakarta
servlet filters, can be added to the Zeebe Gateway.

REST API filters can be added to the Zeebe Gateway with
the same mechanism used for gRPC interceptors. As a
result, the documentaion is very similar, but may
diverge as the Camunda 8 REST API develops futher.
This is why there are two separate documentation
sections for filters and interceptors even though
the content is very similar.

* feat: link to rest api filters in relevant sections of the docs

Link to the REST API filters documentation from
doc pages where gRPC interceptors as mentioned
as well since interceptors and filters perform
the same functionality for the REST and gRPC
parts of the Zeebe Gateway.

* refactor: fix links and code examples for filters

Clean up the REST API filter section of the docs by
setting the correct links to the Jakarta servlet
filter documentation, removing redundant text and
fixing internal links in the Camunda docs.

Also add cross-linking between the interceptor and
filter docs so users are aware of both concepts when
adding custom middleware to their Zeebe installation.

* feat: add filter configuration docs

Expand the gateway configuration docs with a section on
how to configure REST API filters.

* style(formatting): grammatical review

* feat: bump jdk versions in filter docs

Align the JDK versions mentioned in the REST API filter
docs with the supported JDK version for Zeebe.

* feat: bump jdk versions in interceptor docs

Align the JDK versions mentioned in the gRPC Interceptor docs with
the supported JDK version for Zeebe.

---------

Co-authored-by: Christina Ausley <christina.ausley@camunda.com>