Document instance migration phases 3 and 4 (#4348)
* docs: add new instance migration grpc error cases

With 8.6, instance migration will have a few more error cases to ensure
process instances are always migrated correctly.

* docs: add new instance migration limitations

With 8.6, instance migration will have a few more limitations to ensure
process instances are always migrated correctly.

* docs: lift no-mapped boundary events limitation

Boundary events can now be mapped in instance migrations. This results
in migrating the associated subscription.

While this change adjusts the limitations and notes that this mapping
was not supported, it does not yet provide an example to showcase the
behavior.

* wip: add section for dealing with catch events

This is not finished yet, but adds a first section for dealing with
catch events.

* docs: document migrating catch events

* docs: enrich the section for handling catch events

* docs: move the section for handling catch events up

* docs: add supported bpmn elements section

This section can also be moved to another page

* docs: change supported elements headers

These headers should all be H3 under the H2 header for supported
elements. This ensures they're displayed as part of rather than their
own sections.

* docs: move supported elements to limitations

The concept page should focus on what users can achieve with this
feature (the concept) and how to use it. The limitations are in the
bottom so it doesn't distract from teaching the reader about the
concept. Which elements are supported is just further clarifying the
limitations and capabilities.

* docs: move supported elements under limitations

Move the headers so it's part of rather than separate sections.

* docs: add first set of visuals for handling catch events

* docs: update image to the correct one

* docs: add more visuals for handling catch events

* docs: add more visuals for handling catch events

* docs: use event subprocess visual for add/remove catch event

* docs: fix links to the support bpmn elements

* docs: add link to REST API

* docs: update same message name tip

* docs: add example for boundary event detachment

* docs: simplify limitations section

* docs: add another catch event limitation

* docs: use existing bpmn symbol SVGs

* docs: move changing the process instance state discussion to intro

* docs: add supported bpmn elements to limitations section

* docs: remove old comment

* docs: remove excessive sentence

* docs: remove unnecessary section

* docs: emphasize catch event limitation in general

* docs: apply suggestions from code review

* docs: add limitations for multi-instance body

* docs: add operational semantics

* docs: clarify what happens if migration fails

* docs: simplify catch event behaviour explanations

* docs: simplify internal execution explanations

* docs: group limitations section

* docs: add more rejection cases to RPC docs

* docs: add message subscription limitation

* docs: apply suggestions from tech writer review

---------

Co-authored-by: Nico Korthout <korthout@users.noreply.github.com>