fix: add identity request management in gateway
Document the Identity request management feature in
the Zeebe Gateway. Since the configuration properties
for this feature are experimental, we are not documenting
them directly, but onlt linking to the YAML configuration
templates in the camunda/camunda repository. This follows
the same documentation pattern as the Broker's experimental
configuration properties.