Add identity request management in gateway (#4547)
* fix: add identity request management in gateway

Document the Identity request management feature in
the Zeebe Gateway. Since the configuration properties
for this feature are experimental, we are not documenting
them directly, but onlt linking to the YAML configuration
templates in the camunda/camunda repository. This follows
the same documentation pattern as the Broker's experimental
configuration properties.

* adjust note

* Update versioned_docs/version-8.4/self-managed/zeebe-deployment/configuration/gateway.md

Co-authored-by: Sebastian Bathke <sebastian.bathke@camunda.com>

---------

Co-authored-by: Cole Garbo <cole.garbo@camunda.com>
Co-authored-by: Amara Graham <akeller@users.noreply.github.com>
Co-authored-by: Sebastian Bathke <sebastian.bathke@camunda.com>