User Task Listeners update guide changes (#4804)
* docs: describe changed user task events

This is relevant to users updating from 8.6 to 8.7 that rely on the
`ASSIGNING` record from the exporters specifically. Either because they
build their own exporter, or because they use the data in ES/OS or some
other external data storage that has received this information.

* docs: describe assignee not in created events

This is relevant to anyone depending on the `assignee` in the exported
`USER_TASK` records `CREATING/CREATED`.

Co-authored-by: Dmitriy Melnychuk <123628255+ce-dmelnych@users.noreply.github.com>
Co-authored-by: Ana <148554252+ana-vinogradova-camunda@users.noreply.github.com>