Clarify scope of `Execution listeners` in concept page (#4852)
* docs: clarify scope of ELs in `Define an execution listener` section

Updated the `Define an execution listener` section in the `Execution listeners` concept page
to explicitly state that ELs can be configured not only for individual BPMN elements,
but also for the process itself and subprocesses.

This change ensures the documentation clearly highlights the supported scope of execution listeners,
addressing potential ambiguity and improving usability for developers configuring listeners at different levels.

* docs: clarify scope of ELs in `Define an execution listener` section (8.6 version)

Backported updates to the `Define an execution listener` section
in the `Execution listeners` concept page for version 8.6.