Add basic auth to Java client / Spring SDK (#5139)
* fix: correct spring-sdk basic auth configuration

The configuration listed username/password support. This support was dropped, but now it's back!
I only had to modify the mode, the username and password are still provided in the same way.

* feat: introduce java client authentication page

This page will list the different authentication methods that can be used with the java client.
Currently, oidc and x509 are listed under the quick reference. These will be moved to this page in a later commit.

* feat: add basic auth to java client authentication

Adds a section which explains how to configure the java client to work with basic auth.

* feat: add oidc to java client authentication

Adds a section which explains how to configure the java client to work with oidc.

* refactor: move x509 to java client authentication

Moves the existing x509 documentation to the authentication page.

* refactor: remove authentication related docs from the quick reference

Authentication configurations were moved to a separate page. We can remove the details here and reference the new page instead.

* feat: add bash syntax highlighting

We use ```bash``` code blocks in several places. This syntax is not highlighted properly.
By adding bash to the list of additional language we gain this highlighting, making the code blocks easier to read.

* fix: use the declared variable

We created a credentialsProvider, but didn't actually use it.

* refactor: remove applyEnvironmentOverrides flag from code example

I've removed this from the code example and explained how to use it in the text later when we talk about environment variables.

* Adjust tabs

---------

Co-authored-by: Cole Garbo <cole.garbo@camunda.com>