docs: clarify event triggers
It's important for users to understand when the events are triggered
that their listener reacts to.

We can keep this explanation very basic (assigning is triggered when the
user task is being assigned) or we can be explicit about the different
ways to trigger it.

I believe it is more useful for users to know exactly how to trigger the
event. This helps users understand the system they are using.

This comes at the cost of maintenance. For example, we may eventually
add a way for users to update the `priority` of a user task directly
from Tasklist. This is likely implemented by Tasklist calling the update
user task API, and thus automatically triggers the event that the
listener reacts to. We may forget to document this at that time. I would
rather take this risk over harder to understand documentation.

I've also taken the liberty to link the triggers directly to the
relevant documentation. This should make it easier for users to figure
out how to trigger the event.

I've chosen to use a dash (-) rather than repeating the event, as I
think this makes it easier to find all triggers for a single event.