Document process instance migration phase 5 (#5041)
* docs: document escalation event migration

Only supported BPMN elements section is updated because escalation
events can be migrated as any other event type.

* docs: document compensation event migration

Supported BPMN elements section is updated as well as some details are added
related to compensation event creation in the target. Because, the behaviour
of compensation subscription after the migration depends on the state of the
activity that is going to be compensated.

* docs: document joining gateway migration

* docs: rollback 8.7 docs update

* refactor: apply suggestions from review

Co-authored-by: Nico Korthout <korthout@users.noreply.github.com>

* docs: update compensation event migration intro

* docs: update compensation event migration graphics and explanation

* refactor: apply suggestions

Co-authored-by: Nico Korthout <korthout@users.noreply.github.com>

* tw reciew

---------

Co-authored-by: Nico Korthout <korthout@users.noreply.github.com>
Co-authored-by: Cole Garbo <cole.garbo@camunda.com>