Document multi-tenancy for new identity (8.8) (#5327)
* docs: update managing tenants page to new identity

Updates the page with new screenshots and some small textual changes.

* docs: update tenant id constraints

Tenants must follow this regex ^[a-zA-Z0-9_@.-]+$ and have a maximum length of 256 characters.

* docs: remove irrelevant section

There is no longer a way to integrate an external tenant managing component, as tenants are now manages within Zeebe directly.

* docs: correct multi-tenancy configurations

The configurations for multi-tenancy changed in 8.8. Instead of having to configure it for all components we now have a single global config. This simplifies this page a lot.

* refactor: put helm and env vars in tabs

Split the different configurations in different tabs to make the page less cluttered.

* fix: re-add groups to multi tenancy assignment

I removed this, but it's not correct. Groups can still be assigned to a tenant.

* fix: simplify sentence

* fix: remove dash form environment variable

* docs: add multi-tenancy configuration

Adds the option to enable/disable multi-tenancy to the list of configuration options.

* refactor: sort configurations in alphabetical order

* docs: remove outdated multi-tenancy related sections from component pages

Since we moved to a global config these components no longer need to describe how to configure multi-tenancy individually.