Document `updating` user task listener (#5231)
* docs: add updating event as supported

* docs: clarify event triggers

It's important for users to understand when the events are triggered
that their listener reacts to.

We can keep this explanation very basic (assigning is triggered when the
user task is being assigned) or we can be explicit about the different
ways to trigger it.

I believe it is more useful for users to know exactly how to trigger the
event. This helps users understand the system they are using.

This comes at the cost of maintenance. For example, we may eventually
add a way for users to update the `priority` of a user task directly
from Tasklist. This is likely implemented by Tasklist calling the update
user task API, and thus automatically triggers the event that the
listener reacts to. We may forget to document this at that time. I would
rather take this risk over harder to understand documentation.

I've also taken the liberty to link the triggers directly to the
relevant documentation. This should make it easier for users to figure
out how to trigger the event.

I've chosen to use a dash (-) rather than repeating the event, as I
think this makes it easier to find all triggers for a single event.

* docs: use attribute over property

There is no clear difference between attributes or properties. However,
there's a good reason to use attribute over property when discussing the
user task data (e.g. assignee) specifically.

We'll add a new header `changedAttributes` that informs the job worker
about the user task data that changed with the event. Additionally, it's
also the more commonly used term in our Update user task API description
(although that also mentions property once).

Finally, it also helps distinguish these attributes from the user task
listener properties (type/retries).

* docs: describe changed attributes

There was no good place within the table to add this description. So
instead, I've opted for a new section below. We could try something
similar for action.

Generally the description is straightforward. One exception has to be
made regarding the data corrections. A note seems fitting.

* docs: align table header alignment

The other tables in this page all have left-aligned headers. Now this
one as well.

* Fix links

---------

Co-authored-by: Mark Sellings <mark.sellings@camunda.com>