docs: update self-managed update guide (#5430)
In order to pass the initial `assignee` from the user task creation
through the `creating` listeners to the `assigning` event,
we have to store it for later usage. We can only store data in state
when we write it in events. The decision was made to add it to the `creating` event.
To avoid the user task being created with an assignee, it should not be
present in the `created` event.