Describe banned process instance concept (#5736)
* docs: describe banned process instances

I chose to place this under the technical-concepts of Zeebe in the
internal processing section because it's very specific to Zeebe (at this
time). For example, Operate has no way to show banned instances as
different, nor does the orchestration cluster's REST API provide info
about banned instances. Whether this will change is unclear at this
time.

In this page, it fits nicely under the driving the engine section,
because the banned process instance functionality is specific to the
workflow engine (not the stream processor in general).

* TW edits

* docs: highlight banned instance meaning

It's important for users to understand what it means that their process
instance is banned. With this change, we can support that while still
focusing the text on internal processing (context of the page).

---------

Co-authored-by: mesellings <mark.sellings@camunda.com>