docs: describe the assignee correction intricacies
The rule that you can only correct the assignee on the creating event if
and only if the assignee is not already specified by the process
requires us to help the user a bit.

The rule exists because without it, we would not know what do. For
example, consider a user task where the process specifies 'foo' as
assignee. The creating listener triggers, and the listener corrects the
assignee from `foo` to `bar`. However, the user task is never `CREATED`
with an assignee, instead the assignee is passed along to the
`ASSIGNING` event which triggers next. Now the following issues exists:
- if we pass `bar` to `assigning`, then shouldn't we do so as well when
  no assignee was specified? This would break with the rule that
  corrections don't trigger other lifecycle events.
- if we pass `foo` to `assigning`, then the correction is ignored which
  would be surprising to the user.

Instead, we simply disallow this. We therefore need to explain to users
what they should do instead:
- you can change the assignee in the assigning event
- if none is specified by the process, then you can use the creating
  event to set one instead