Document `creating` and `canceling` user task listeners (#5790)
* docs: support creating and canceling events

* docs: add creating and canceling triggers

* docs: describe user task lifecycle

This makes it easier to understand where listeners actually come into
play.

* docs: move supported events to about section

In the user task lifecycle section we can go into more detail about the
different events, but it's useful to have this quick overview of
supported user task listener events in the About user task listeners
section.

* docs: add heading for event triggers

This further clarifies how listeners can be used, because the events
need to be triggered

* docs: rearrange sections

The docs should teach new users how the concept works, and allow
advanced users to quickly look up things from reference.

I've rearranged this page a bit, to accommodate that.

First you learn about user task listeners in the about section.
Users that already have a process, may then want to trigger a user task
listener.
If they don't yet have a process with a user task listener, they may
want to define a user task listener.
Lastly, they may want to implement the task listener

* docs: clarify that not all events can be denied

* docs: remove limited events support limitation

We now support all user task events

* docs: elaborate on canceling triggers

A user task can be canceled in two ways:
- termination of the task due to process instance cancelation
- termination of the task due to triggered interrupting catch event

* docs: link task listeners concept in update guide

Initially, I wanted to make a change to the update guide: document that
assignee is filled in the CREATING event. I discovered that no change
was needed. However, while reading the page I noticed the missing link.

* docs: link user tasks from task listeners page

* docs: describe the assignee correction intricacies

The rule that you can only correct the assignee on the creating event if
and only if the assignee is not already specified by the process
requires us to help the user a bit.

The rule exists because without it, we would not know what do. For
example, consider a user task where the process specifies 'foo' as
assignee. The creating listener triggers, and the listener corrects the
assignee from `foo` to `bar`. However, the user task is never `CREATED`
with an assignee, instead the assignee is passed along to the
`ASSIGNING` event which triggers next. Now the following issues exists:
- if we pass `bar` to `assigning`, then shouldn't we do so as well when
  no assignee was specified? This would break with the rule that
  corrections don't trigger other lifecycle events.
- if we pass `foo` to `assigning`, then the correction is ignored which
  would be surprising to the user.

Instead, we simply disallow this. We therefore need to explain to users
what they should do instead:
- you can change the assignee in the assigning event
- if none is specified by the process, then you can use the creating
  event to set one instead

* docs: describe user task headers and data

User task data is merged into the user task's task headers. This should
be documented for clarity.

* docs: link to catch events

Might be useful for users to learn if they don't know it yet

* docs: fix typo

Co-authored-by: Dmitriy Melnychuk <123628255+ce-dmelnych@users.noreply.github.com>

* docs: expand supported event types definition

It's useful to document the supported values in this section, for
readers that are only interested in how a listener can be defined.
However, we must be complete, so we now need to specify all 5 supported
event types.

* style(formatting): grammatical tweaks

* docs: clarify correcting assignee on creating restriction

Co-authored-by: Christina Ausley <84338309+christinaausley@users.noreply.github.com>

---------

Co-authored-by: Dmitriy Melnychuk <123628255+ce-dmelnych@users.noreply.github.com>
Co-authored-by: Christina Ausley <christina.ausley@camunda.com>
Co-authored-by: Christina Ausley <84338309+christinaausley@users.noreply.github.com>