Document processing metrics availability
This came up in a support case SUPPORT-27804, where it was not clear to
the customer that in order to see the
zeebe_process_instance_creations_total or the
zeebe_incident_events_total, there needs to be some activity on the
cluster that causes the recording of these metric events.

To avoid this in the future, I've added a note to the page in the
Available metrics section.