fix(self-managed): align port numbers
As part of iteration 1 of

https://github.com/camunda/team-distribution/issues/578 ,

detailed in this comment
https://github.com/camunda/team-distribution/issues/578#issuecomment-3162202540

I will be aligning the port numbers between each deployment method.

This PR updates all documentation recommendations for local
port-forwards to match what is recommended by each of the distributions
(compose, helm, c8run) so that the same port can be recognized
regardless of which distribution you're on.