fix(docs): add notice about change in spring zeebe sdk
When the Spring Zeebe SDK was released with 8.5.0, it had some
differences to the Spring Zeebe Client that it was forked on. In
particular, there was one breaking change.

We now document that breaking change in the release notes for any
customers updating from Spring-Zeebe Client (>=8.5.3) to the Spring
Zeebe SDK.

Note that the breaking change originated from a bug fix which was also
patched in Spring-Zeebe Client 8.5.4.