fix: do not invoke scrollToHash if no diagrams are present in the page (#6758)
# Cause
The function renderDiagrams would call scrollToHash on every render even if no bpmn diagrams are present in the page.
The function scrollToHash would throw an error if the location.hash starts with a number like '#1-example'.

# Fix
If no diagrams are present in the page, renderDiagrams now just return immediately instead.