refactor: drop unused ports
The ports 26500 and 18080 do not need to be mapped from the container of
the KIND cluster to the host because these ports are mapped via Ingress.
These 2 ports are unnecessarily mapped, which means that it is
problematic if you were running a KIND cluster as well as an instance of
C8Run.

Signed-off-by: Jesse Simpson <jesse.simpson@camunda.com>